added a list of popular places in the selected category added the function of finding 10 top places for the last month

+++ src/main/java/com/netcracker/edu/review/service/ReviewService.java
@@ -1,5 +1,8 @@
 package com.netcracker.edu.review.service;
 
+import com.netcracker.edu.review.error.ReviewByRatingNotFoundException;
+import com.netcracker.edu.review.error.ReviewNotFoundException;
+import com.netcracker.edu.review.model.Mark;
 import com.netcracker.edu.review.model.Review;
 import com.netcracker.edu.review.model.ui.UiReview;
 import com.netcracker.edu.review.repository.ReviewRepository;
@@ -9,7 +12,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import  com.netcracker.edu.review.model.Mark;
+
 import java.util.Date;
 import java.util.List;
 
@@ -22,43 +25,64 @@ public class ReviewService {
     @Autowired
     private MarkService markService;
 
-    @Value("${countOfReviews.size-of-page}")
+
+    @Autowired
+    private RatingService ratingService;
+
+    @Value("${count-of-reviews.size-of-page}")
     public int sizeOfPage;
 
-    @Value("${countOfReviews.date-creation}")
+    @Value("${count-of-reviews.date-creation}")
+
     public String dateOfCreation;
 
-    public Review createReview(UiReview uiReview, Mark mark) {
+    public Review createReview(UiReview uiReview) {
 
-        mark = markService.findMarkByValue(uiReview.getMark());
+        Mark mark = markService.findMarkByValue(uiReview.getMark());
 
+        if (uiReview.getMark() != null) {
+            ratingService.setRating(uiReview, mark);
+        }
+        ratingService.sortTenList();
         return reviewRepository.saveAndFlush(new Review(uiReview, mark));
     }
 
-    public List<Review> findReviewByAuthorId(int authorId,int page) {
-        Pageable pageable = PageRequest.of(page, sizeOfPage, Sort.by(dateOfCreation).descending());
 
-      return reviewRepository.findReviewByAuthorId(authorId, pageable);}
+    public List<Review> findReviewByAuthorId(int authorId, int page) {
+        Pageable pageable = PageRequest.of(page, sizeOfPage, Sort.by(dateOfCreation).descending());
+        List<Review> review = reviewRepository.findReviewByAuthorId(authorId, pageable);
+        if (review.isEmpty()) {
+            throw new ReviewNotFoundException();
+        } else return review;
+    }
+
 
     public List<Review> findReviewByPlaceId(int placeId, int page) {
-        Pageable pageable = PageRequest.of(page, sizeOfPage,  Sort.by(dateOfCreation).descending());
-        return reviewRepository.findReviewByPlaceId(placeId, pageable);
+        Pageable pageable = PageRequest.of(page, sizeOfPage, Sort.by(dateOfCreation).descending());
+        List<Review> reviews = reviewRepository.findReviewByPlaceId(placeId, pageable);
+        if (reviews.isEmpty()) {
+            throw new ReviewNotFoundException();
+        } else return reviews;
     }
 
-    public List<Review> findReviewByPlaceIdAndAuthorId(int placeId,int authorId, int page) {
-        Pageable pageable = PageRequest.of(page, sizeOfPage,  Sort.by(dateOfCreation).descending());
-        return reviewRepository.findReviewByPlaceIdAndAuthorId(placeId, authorId, pageable);
+    public List<Review> findReviewByPlaceIdandAuthorId(int placeId, int authorId, int page) {
+        Pageable pageable = PageRequest.of(page, sizeOfPage, Sort.by(dateOfCreation).descending());
+        List<Review> review = reviewRepository.findReviewByPlaceIdAndAuthorId(placeId, authorId, pageable);
+        if (review.isEmpty()) {
+            throw new ReviewNotFoundException();
+        } else return review;
     }
 
-    public Review updateReviewById(int id, UiReview uiReviewUpdated, Mark mark) {
+    public Review updateReviewById(int id, UiReview uiReviewUpdated) {
 
-        mark = markService.findMarkByValue(uiReviewUpdated.getMark());
+        Mark mark = markService.findMarkByValue(uiReviewUpdated.getMark());
 
         Review review = reviewRepository.getById(id);
         if (uiReviewUpdated.getReview() != null) {
             review.setReview(uiReviewUpdated.getReview());
         }
         if (uiReviewUpdated.getMark() != null) {
+            ratingService.updateRatingWithReview(review, mark);
             review.setMark(mark);
         }
         review.setDateEdit(new Date());
@@ -66,9 +90,24 @@ public class ReviewService {
 
     }
 
-    public void deleteReviewByid(int id) {
-        reviewRepository.deleteById(id);
+    public void deleteReviewById(int id) {
+        if (!reviewRepository.findById(id).isPresent()) {
+            throw new ReviewNotFoundException();
+        } else {
+            Review review = reviewRepository.findById(id).get();
+            ratingService.deleteRatingWithReview(review);
+            reviewRepository.deleteById(id);
         }
+    }
+
+    public List<Review> findReviewByMarkId(int markId, int page) {
+        Pageable pageable = PageRequest.of(page, sizeOfPage, Sort.by(dateOfCreation).descending());
+        List<Review> review = reviewRepository.findReviewByMark_Id(markId, pageable);
+        if (review.isEmpty()) {
+            throw new ReviewByRatingNotFoundException(Integer.toString(markId));
+        } else return review;
+    }
+
 }
